Fixed tower moving.
Made a more general (and reusable) moveTower method
in Inventory to make sure changes on screen are
reflected in the Inventory while the Inventory is
still agnostic about which button index the Tower is
assigned to.

+++ src/main/java/seng201/team56/models/Player.java
@@ -9,7 +9,7 @@ public class Player {
     private final Inventory inventory;
     private Difficulty difficulty;
     private final int maxRounds;
-    private double money;
+    private int money;
 
     public int getMaxRounds() {
         return maxRounds;
@@ -66,7 +66,7 @@ public class Player {
      * Getter for how much money the player has
      * @return money
      */
-    public double getMoney() {
+    public int getMoney() {
         return money;
     }
 
@@ -74,7 +74,7 @@ public class Player {
      * Add money to the player
      * @param amount the amount to add
      */
-    public void addMoney(double amount) {
+    public void addMoney(int amount) {
         money += amount;
     }
